// facilita-motors-gest-o-completa-main/src/types/database.ts
export interface Cliente {
  id: string;
  nome: string;
  cpf_cnpj: string;
  telefone: string;
  email: string | null;
  endereco_cep: string | null;
  endereco_rua: string | null;
  endereco_numero: string | null;
  endereco_bairro: string | null;
  endereco_cidade: string | null;
  endereco_estado: string | null;
  data_nascimento: string | null;
  criado_em: string | null;
  atualizado_em: string | null;
}

export interface Veiculo {
  id: string;
  cliente_id: string;
  marca: string;
  modelo: string;
  ano: number | null;
  cor: string | null;
  placa: string;
  quilometragem: number | null;
  observacoes: string | null;
  criado_em: string | null;
}